+++ spec/datadog/appsec/waf/context_spec.rb
@@ -43,9 +43,9 @@ RSpec.describe Datadog::AppSec::WAF::Context do
         expect(result).not_to be_timeout
         expect(result.status).to eq(:ok)
         expect(result.events).to eq([])
-        expect(result.total_runtime).to be >= 0
+        expect(result.duration).to be >= 0
         expect(result.actions).to eq({})
-        expect(result.derivatives).to eq({})
+        expect(result.attributes).to eq({})
       end
     end
 
@@ -56,9 +56,9 @@ RSpec.describe Datadog::AppSec::WAF::Context do
         expect(result).not_to be_timeout
         expect(result.status).to eq(:ok)
         expect(result.events).to eq([])
-        expect(result.total_runtime).to be >= 0
+        expect(result.duration).to be >= 0
         expect(result.actions).to eq({})
-        expect(result.derivatives).to eq({})
+        expect(result.attributes).to eq({})
       end
     end
 
@@ -69,9 +69,9 @@ RSpec.describe Datadog::AppSec::WAF::Context do
         expect(result).not_to be_timeout
         expect(result.status).to eq(:match)
         expect(result.events).to match_array([{"rule" => anything, "rule_matches" => anything}])
-        expect(result.total_runtime).to be >= 0
+        expect(result.duration).to be >= 0
         expect(result.actions).to eq({"block_request" => {"grpc_status_code" => "10", "status_code" => "403", "type" => "auto"}})
-        expect(result.derivatives).to eq({})
+        expect(result.attributes).to eq({})
       end
     end
 
@@ -82,9 +82,9 @@ RSpec.describe Datadog::AppSec::WAF::Context do
         expect(result).not_to be_timeout
         expect(result.status).to eq(:match)
         expect(result.events).to match_array([{"rule" => anything, "rule_matches" => anything}])
-        expect(result.total_runtime).to be >= 0
+        expect(result.duration).to be >= 0
         expect(result.actions).to eq({"block_request" => {"grpc_status_code" => "10", "status_code" => "403", "type" => "auto"}})
-        expect(result.derivatives).to eq({})
+        expect(result.attributes).to eq({})
       end
     end
 
@@ -188,7 +188,7 @@ RSpec.describe Datadog::AppSec::WAF::Context do
       end
 
       context "with schema extraction" do
-        it "populates derivatives" do
+        it "populates attributes" do
           waf_args = {
             "server.request.query" => {
               "hello" => "EMBED"
@@ -202,13 +202,13 @@ RSpec.describe Datadog::AppSec::WAF::Context do
 
           aggregate_failures("result") do
             expect(result.status).to eq :ok
-            expect(result.derivatives).to eq({"_dd.appsec.s.req.query" => [{"hello" => [8]}]})
+            expect(result.attributes).to eq({"_dd.appsec.s.req.query" => [{"hello" => [8]}]})
           end
         end
       end
 
       context "without schema extraction" do
-        it "populates derivatives" do
+        it "populates attributes" do
           waf_args = {
             "server.request.query" => {
               "hello" => "EMBED"
@@ -222,7 +222,7 @@ RSpec.describe Datadog::AppSec::WAF::Context do
 
           aggregate_failures("result") do
             expect(result.status).to eq :ok
-            expect(result.derivatives).to be_empty
+            expect(result.attributes).to be_empty
           end
         end
       end