+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 
 - Change `LibDDWAF::Object#input_truncated?` to `LibDDWAF::Object#truncated?` method
 - Change `LibDDWAF::Object#mark_as_input_truncated?` to `LibDDWAF::Object#mark_truncated?` method
-- Change `Result` class initialization interface
 - Change `Result#timeout` to `Result#timeout?` method
+- Change `Result#derivatives` to `Result#attributes` method
+- Change `Result#total_runtime` to `Result#duration` method
 
 # 2025-09-02 v1.24.1.2.0